integrated paypal and added some validations

+++ products/cart.php
@@ -2,29 +2,29 @@
 <?php require "../config/config.php"; ?>
 <?php
 
-	if (!isset($_SESSION['user_id'])) {
-		header("location: " . APPURL . "");
-	}
+if (!isset($_SESSION['user_id'])) {
+	header("location: " . APPURL . "");
+}
 
-	$products = $conn->query("SELECT * FROM cart WHERE user_id = '$_SESSION[user_id]'");
-	$products->execute();
+$products = $conn->query("SELECT * FROM cart WHERE user_id = '$_SESSION[user_id]'");
+$products->execute();
 
-	$allProducts = $products->fetchAll(PDO::FETCH_OBJ);
+$allProducts = $products->fetchAll(PDO::FETCH_OBJ);
 
 
-	// Cart Total
-	$cartTotal = $conn->query("SELECT SUM(quantity * price) AS total FROM cart WHERE user_id = '$_SESSION[user_id]'");
-	$cartTotal->execute();
+// Cart Total
+$cartTotal = $conn->query("SELECT SUM(quantity * price) AS total FROM cart WHERE user_id = '$_SESSION[user_id]'");
+$cartTotal->execute();
 
-	$allCartTotal = $cartTotal->fetch(PDO::FETCH_OBJ);
+$allCartTotal = $cartTotal->fetch(PDO::FETCH_OBJ);
 
-	// Proceed To checkout
-	if(isset($_POST['checkout'])) {
+// Proceed To checkout
+if (isset($_POST['checkout'])) {
 
-		$_SESSION['total_price'] = $_POST['total_price'];
+	$_SESSION['total_price'] = $_POST['total_price'];
 
-		header("location: checkout.php");
-	}
+	header("location: checkout.php");
+}
 
 ?>
 
@@ -50,44 +50,48 @@
 		<div class="row">
 			<div class="col-md-12 ftco-animate">
 				<div class="cart-list">
-					<table class="table">
-						<thead class="thead-primary">
-							<tr class="text-center">
-								<th>&nbsp;</th>
-								<th>&nbsp;</th>
-								<th>Product</th>
-								<th>Price</th>
-								<th>Quantity</th>
-								<th>Total</th>
-							</tr>
-						</thead>
-						<tbody>
-							<?php foreach ($allProducts as $product) : ?>
+					<?php if (count($allProducts) > 0) : ?>
+						<table class="table">
+							<thead class="thead-primary">
 								<tr class="text-center">
-									<td class="product-remove"><a href="delete-product.php?id=<?php echo $product->id; ?>"><span class="icon-close"></span></a></td>
+									<th>&nbsp;</th>
+									<th>&nbsp;</th>
+									<th>Product</th>
+									<th>Price</th>
+									<th>Quantity</th>
+									<th>Total</th>
+								</tr>
+							</thead>
+							<tbody>
+								<?php foreach ($allProducts as $product) : ?>
+									<tr class="text-center">
+										<td class="product-remove"><a href="delete-product.php?id=<?php echo $product->id; ?>"><span class="icon-close"></span></a></td>
 
-									<td class="image-prod">
-										<div class="img" style="background-image:url(<?php echo APPURL ?>/images/<?php echo $product->image; ?>);"></div>
-									</td>
+										<td class="image-prod">
+											<div class="img" style="background-image:url(<?php echo APPURL ?>/images/<?php echo $product->image; ?>);"></div>
+										</td>
 
-									<td class="product-name">
-										<h3><?php echo $product->name; ?></h3>
-										<p><?php echo $product->description; ?></p>
-									</td>
+										<td class="product-name">
+											<h3><?php echo $product->name; ?></h3>
+											<p><?php echo $product->description; ?></p>
+										</td>
 
-									<td class="price">$<?php echo $product->price; ?></td>
+										<td class="price">$<?php echo $product->price; ?></td>
 
-									<td>
-										<div class="input-group mb-3">
-											<input disabled type="text" name="quantity" class="quantity form-control input-number" value="<?php echo $product->quantity; ?>" min="1" max="100">
-										</div>
-									</td>
+										<td>
+											<div class="input-group mb-3">
+												<input disabled type="text" name="quantity" class="quantity form-control input-number" value="<?php echo $product->quantity; ?>" min="1" max="100">
+											</div>
+										</td>
 
-									<td class="total">$<?php echo $product->price * $product->quantity; ?></td>
-								<?php endforeach; ?>
-								</tr><!-- END TR-->
-						</tbody>
-					</table>
+										<td class="total">$<?php echo $product->price * $product->quantity; ?></td>
+									<?php endforeach; ?>
+									</tr><!-- END TR-->
+							</tbody>
+						</table>
+					<?php else : ?>
+						<p>Cart Empty</p>
+					<?php endif; ?>
 				</div>
 			</div>
 		</div>
@@ -110,13 +114,16 @@
 					<hr>
 					<p class="d-flex total-price">
 						<span>Total</span>
+						<?php if($allCartTotal->total > 0) : ?>
 						<span>$<?php echo $allCartTotal->total + 10 - 3; ?></span>
+						<?php endif; ?>
 					</p>
 				</div>
 				<form action="cart.php" method="POST">
 					<input type="hidden" name="total_price" value="<?php echo $allCartTotal->total + 10 - 3; ?>">
-
+						<?php if($allCartTotal->total > 0) :?>
 					<button name="checkout" type="submit" class="btn btn-primary py-3 px-4">Proceed to Checkout</button>
+					<?php endif; ?>
 				</form>
 			</div>
 		</div>
